Library/Mobile Documents/com~apple~CloudDocs/AriesView/DATA-AI/ocr/requirements.txt
fastapi
Pillow
PyMuPDF
pydantic
httpx
python-multipart
uvicorn[standard]
python-dotenv

pypdf
pdf2image
tqdm
paddleocr
opencv-python-headless
dotenv
